COMP: Make `cvxpy` an optional dependency.
Also remove import aliases from `dwd/__init__.py` so that, if `cvxpy` is not installed, no `ImportError` is raised. This requires users to explicitly `import dwd.socp_dwd` to use the conic solver.

Renamed `dwd.dwd` to `dwd.socp_dwd` to clarify the solver method used, and be more consistent with the other solvers.